update labels from deep sprint planning
added the delegate and decide label.

I added these to the middle of the stack; not sure if that will change the IDs of all of them and mess things up royally and if this should be added to the bottom; but we'll see! (also not sure how well reverting works for IaC...)