Seeking clarity on pull requests
DNS pull requests seem like they are beyond the scope of OpsRotation and should be handled by someone with the appropriate knowledge, skills and abilities. I added a note to let future OpsRotation folks know that those pull requests that are in the board do not have to be reviewed as a part of OpsRotation.