more specific about which GitHub Repos Admins
As someone who is new to the TTS Tech Portfolio, I would like to be made Admin to the GitHub Organizations that have already allowed others admin privileges so that I can backfill for them when needed. 

As someone who is supporting the TTS Tech Portfolio, I would like to be made Admin to the GitHub Organizations [1] that are under the TTS umbrella so that I can properly triage situations as they occur and perform functions that are within our responsibility. 

[1]